pyproject.toml
[project]
name = "ibkr-trader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ib-insync>=0.9.86",
    "pandas>=3.0.0",
    "python-dotenv>=1.2.1",
]
